Add JSONB documentation to the datatype list in YCQL docs.
Summary:
The JSONB link on our datatypes page was broken since we didn't have associated
documentation for this type similar to our other types (we only described it in the explore
section).

I've added documentation for the JSONB type explaining the syntax, semantics and various operators.

Test Plan: view locally.

Reviewers: kannan, karthik, sid

Reviewed By: sid

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D5059