ENG-4162: Adding a ulimit backreference in the kubernetes tab
Summary:
Seems that ulimits need to be setup on the actual k8s nodes, so we should add these to the
docs. Since the initial checklist forward references the manual deployment section, I had the helm
prereq section backreference to that as well.

Perhaps a slightly better option would be that checklist has everything, including ntp and ulimits
and then both manual and kubernetes backreference to that, which is the very first page.

Test Plan: http://bogdan-dev-server.c.yugabyte.internal:1313/latest/deploy/kubernetes/helm-chart/#prerequisites

Reviewers: sid, ram, ravi

Reviewed By: ravi

Differential Revision: https://phabricator.dev.yugabyte.com/D5789